requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
python-multipart
pickle
python-dotenv
boto3
sklearn